pyproject.toml
[project]
name = "prep-study-assistant"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.5",
    "fastapi>=0.116.1",
    "loguru>=0.7.3",
    "openai>=1.97.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "setuptools>=80.9.0",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
